RLSa2va/requirements.txt
# Sa2VA Requirements
# Python 3.10+

# Core dependencies
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.37.0

# MMEngine ecosystem
mmengine>=0.10.0
mmcv>=2.0.0

# Image processing
opencv-python>=4.8.0
pillow>=10.0.0
scikit-image>=0.21.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Model hub
huggingface_hub>=0.19.0
safetensors>=0.4.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Development
ipython>=8.12.0
jupyter>=1.0.0

# Optional: Acceleration
# accelerate>=0.25.0
# bitsandbytes>=0.41.0

# Optional: Distributed training
# deepspeed>=0.12.0
